service1/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlmodel==0.0.8
strawberry-graphql==0.96.0
boto3==1.26.137
python-multipart==0.0.5
asyncpg==0.27.0